python-expenses/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.10
redis==5.0.1
SQLAlchemy==2.0.23
